refactor: address second batch of ulong-migration review feedback
- State.Flat: make IFlatDbConfig CompactSize/MinReorgDepth/MaxReorgDepth ulong
  (CompactionOffset stays long for its -1 sentinel); ICompactionSchedule.GetCompactSize
  returns ulong; drop the resulting casts in PersistenceManager/SnapshotCompactor and
  add a NextBlockToPersist helper for the duplicated PreGenesis expression. CompactionSchedule
  keeps decoding the stored offset as long so a corrupt negative still regenerates.
- AccountProof: fetch all JSON converters upfront in Write instead of mid-method.
- DifficultyBombDelaysConverter: throw JsonException (not ArgumentException) on bad JSON.
- Ethereum.Test.Base: LegacyTransactionJson GasLimit/Nonce to ulong (ULongConverter handles hex).
- Trie: name the TrieStore to-be-persisted sentinel; use existing int overloads in TreeStoreTests.
- Xdc: MaxSignableBlockPeriods to ulong; named NoRound sentinel in XdcHotStuff; add ulong
  AddBlocks overload; simplify RewardTests header mocks; drop RpcModuleTests casts.
- TxPool tests: use WithNonce(int) overload to drop (ulong) loop casts.
- MemoryHintMan: NettyMemoryEstimator.Estimate returns ulong, dropping estimate casts.
- tools: CMSketch index helper; drop redundant MaxDepth cast in T8nBlockHashProvider.

+++ src/Nethermind/Ethereum.Test.Base/JsonToEthereumTest.cs
@@ -244,9 +244,9 @@ namespace Ethereum.Test.Base
             Transaction transaction = new()
             {
                 Value = transactionJson.Value,
-                GasLimit = (ulong)transactionJson.GasLimit,
+                GasLimit = transactionJson.GasLimit,
                 GasPrice = transactionJson.GasPrice,
-                Nonce = (ulong)transactionJson.Nonce,
+                Nonce = transactionJson.Nonce,
                 To = transactionJson.To,
                 Data = transactionJson.Data,
                 Signature = new Signature(transactionJson.R, transactionJson.S, transactionJson.V)
