+++ src/Nethermind/Nethermind.Db/IFlatDbConfig.cs
@@ -14,7 +14,7 @@ public interface IFlatDbConfig : IConfig
     long CompactionOffset { get; set; }
 
     [ConfigItem(Description = "Compact size", DefaultValue = "32")]
-    int CompactSize { get; set; }
+    ulong CompactSize { get; set; }
 
     [ConfigItem(Description = "Enabled", DefaultValue = "false")]
     bool Enabled { get; set; }
@@ -35,10 +35,10 @@ public interface IFlatDbConfig : IConfig
     int MaxInFlightCompactJob { get; set; }
 
     [ConfigItem(Description = "Max reorg depth", DefaultValue = "256")]
-    int MaxReorgDepth { get; set; }
+    ulong MaxReorgDepth { get; set; }
 
     [ConfigItem(Description = "Minimum reorg depth", DefaultValue = "128")]
-    int MinReorgDepth { get; set; }
+    ulong MinReorgDepth { get; set; }
 
     [ConfigItem(Description = "Lower bound, in bytes, for the RocksDB write buffer (memtable) size of the flat-state columns. The per-batch adjuster never shrinks a column's memtable below this value. Raising it lets frequent small persistence batches (small CompactSize) coalesce and deduplicate in the memtable instead of churning L0, decoupling write amplification from CompactSize.", DefaultValue = "16777216")]
     long PersistenceWriteBufferFloor { get; set; }
