+++ src/Nethermind/Nethermind.State.Flat/ResourcePool.cs
@@ -19,7 +19,7 @@ public class ResourcePool(IFlatDbConfig flatConfig) : IResourcePool
     private readonly Dictionary<Usage, ResourcePoolCategory> _categories = new()
     {
         // For main BlockProcessing once a compacted snapshot is persisted, all `flatConfig.CompactSize` snapshot content will be returned.
-        { Usage.MainBlockProcessing, new ResourcePoolCategory(Usage.MainBlockProcessing, flatConfig.CompactSize + 8, 2) },
+        { Usage.MainBlockProcessing, new ResourcePoolCategory(Usage.MainBlockProcessing, (int)flatConfig.CompactSize + 8, 2) },
 
         // PostMainBlockProcessing is a special usage right after the commit of `MainBlockProcessing` which only commit once and never modified.
         { Usage.PostMainBlockProcessing, new ResourcePoolCategory(Usage.PostMainBlockProcessing, 1, 1) },
